requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.1
django-cors-headers==3.13.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
pycodestyle==2.9.1
PyJWT==2.4.0
python-magic==0.4.27
python-magic-bin==0.4.13
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
